convert community page
